requirements.txt
gradio>=4.0.0
youtube-transcript-api
langchain
langchain-community
langchain-text-splitters
langchain-openai
sentence-transformers
faiss-cpu
python-dotenv
numpy